services/user_server/pyproject.toml
[tool.poetry]
name = "user_server"
version = "0.0.1"
description = ""
authors = ["Новиков Илья <nov.ilucha@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = {extras = ["email"], version = "^2.4.2"}
aerich = "^0.7.2"
tortoise-orm = "^0.20.0"
asyncpg = "^0.28.0"
pydantic-settings = "^2.1.0"
passlib = "^1.7.4"
pydantic-core = "^2.14.5"
bcrypt = "^4.1.1"
orjson = "^3.9.10"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"

[tool.aerich]
tortoise_orm = "src.database.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
